requirements.txt
python-telegram-bot==20.6
together==0.2.5
python-dotenv==1.0.0
